#2245 Experimental redesign of ribbon and footer, WIP
This started as an attempt to incorporate the CGI extranet design
guidelines into #define's UI. It snowballed from there into a redesign
of the header/footer.
Random notes/principals
- The ribbon no longer stays in a fixed place.
- The quick access menu has been broken into seperate menus
- The Settings tab is now 'Manage'. The intention is that only
project managers see this
- There's a new user thumnail/menu - user preferences are accessed from
there.
- The footer contains links that are global to #define
e.g. sales web, community
- The help link in the ribbon becomes a word
it should be context sensitive (not implemented yet)
- !Previous/Next navigation is moved to a predictable location
- Page headings have a unified style
- Design principal: on every page user should be able to copy the page
title and paste it as a wiki link

Ambitions:
- Eliminate toolbars from the ribbon
leaving only the title/breadcrumbs and the tabs/search

git-svn-id: file:///mnt/ramdisk/d4-svn/branches/features/2245-UI-experiment@24166 484df14f-75d2-4b3f-8ba0-490e05069a01